requirements.txt
streamlit==1.55.0
openai==1.109.1
PyMuPDF==1.27.2.2
python-docx==1.2.0
Jinja2==3.1.6
pdfkit==1.0.0
lxml==6.0.2
python-dotenv==1.1.1
